% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R4WML_core.R
\name{from_wml_payload}
\alias{from_wml_payload}
\title{Create a data frame from a Watson Machine Learning result}
\usage{
from_wml_payload(wml_results)
}
\description{
Given a Watson Machine Learning result, return a data frame
}
\examples{
from_wml_payload()
}
\keyword{WML}
\keyword{payload}
