% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R4WML_core.R
\name{get_wml_auth_headers}
\alias{get_wml_auth_headers}
\title{Get Watson Machine Learning (WML) authorization headers}
\usage{
get_wml_auth_headers(wml_url, username, password)
}
\arguments{
\item{username}{The WML user name}

\item{password}{The WML password}

\item{token_url}{The WML URL to autenticate against}
}
\value{
Authorization headers for use in subsequent requests to WML
}
\description{
Given a Watson Machine Learning URL, username and password, return an authorization headers for use in subsequent requests to WML.
}
\examples{
get_wml_auth_headers()
}
\keyword{WML}
\keyword{authorization}
