% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R4WML_core.R
\name{get_wml_auth_token}
\alias{get_wml_auth_token}
\title{Get a Watson Machine Learning (WML) authentication token}
\usage{
get_wml_auth_token(token_url, username, password)
}
\arguments{
\item{token_url}{The WML URL to autenticate against}

\item{username}{The WML user name}

\item{password}{The WML password}
}
\value{
Authentication token for use in subsequent requests to WML
}
\description{
Given a Watson Machine Learning URL, username and password, return an authentication token for use in subsequent requests to WML.
}
\examples{
get_wml_auth_token()
}
\keyword{WML}
\keyword{authentication}
