% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R4WML_core.R
\name{to_wml_payload}
\alias{to_wml_payload}
\title{Create a Watson Machine Learning (WML) payload from an R list}
\usage{
to_wml_payload(df, columns = names(df))
}
\description{
Given an R list, return a JSON payload for sending to a WML scoring endpoint
}
\examples{
to_wml_payload()
}
\keyword{WML}
\keyword{payload}
