% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R4WML_core.R
\name{wml_score}
\alias{wml_score}
\title{Use a Watson Machine Learning REST endpoint to score a payload of records}
\usage{
wml_score(scoring_url, auth_headers, payload, verbose = FALSE)
}
\description{
Given a WML REST endpoint (URL) and JSON payload, score the records in the payload
}
\examples{
wml_score()
}
\keyword{WML}
\keyword{scoring}
