% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4wml_core.R
\name{get_wml_auth_token}
\alias{get_wml_auth_token}
\title{Get a Watson Machine Learning (WML) authentication token}
\usage{
get_wml_auth_token(token_url, username, password)
}
\description{
Given a Watson Machine Learning URL, username and password, return an authentication token for use in subsequent requests to WML.
}
\examples{
get_wml_auth_token()
}
\keyword{WML}
\keyword{authentication}
